requirements.txt
# BentoML and ML libraries
bentoml==1.4.35
numpy==2.4.1
pandas==2.3.3
pyjwt==2.11.0
scikit-learn==1.8.0

# Development / testing
pytest==9.0.2
requests==2.32.5
setuptools==82.0.0